Add files via upload
Program in deployment

Description:
The keylogger uses the OS, Thread and pynput to capture all the keys presed by the customer. All the results will be stored on .txt file for analysis.
The name and path of directory can be modified depending of the use or how easy you want to see the results.